Move schnorr to the crypto package
